Quickfix: load CNN weights from correct GitHub Pages path
For some reason the CNN weight paths are interpreted as relative to the root of
my entire GitHub Pages site (https://mkremins.github.io), rather than relative to
the Blabrecs subdirectory specifically (https://mkremins.github.io/blabrecs).
This commit forces the weights to be loaded from the correct path for the deployed
GitHub Pages version of the site, but uses an incorrect weights path for local
development. The quickfix for local development is to make a local folder called
`blabrecs` and copy the weights file (`group1-shard1of1.bin`) into that folder,
but we really ought to figure out a more principled solution.